Bugs / Improvement + Add new form builder page for public to submit or edit + Create new form + Remove knex + Remove unused page for form builder
