+++ composables/formBuilderList.js
@@ -1,0 +1,5 @@
+import forms from "@@/assets/js/rs-form-builder/list.json";
+
+export default function () {
+  return forms;
+}
